style : renaming folders and adding logs for debugging
1. Logs for checking request lifecycle.
