+++ backend/src/app.module.ts
@@ -6,7 +6,9 @@ import { LoggerMiddleware } from "./middlewares/logger.middleware";
 import { UsersModule } from "./users/users.module";
 import * as path from "path";
 import { Users } from "./users/entities/user.entity";
-import { AuthModule } from './auth/auth.module';
+import { AuthModule } from "./auth/auth.module";
+import { APP_FILTER } from "@nestjs/core";
+import { HttpExceptionFilter } from "./http-exception.filter";
 
 @Module({
   imports: [
