+++ backend/src/middlewares/logger.middleware.ts
@@ -20,6 +20,7 @@ export class LoggerMiddleware implements NestMiddleware {
         `${method} ${originalUrl} ${statusCode} ${contentLength} - ${userAgent} ${ip}`
       );
     });
+    console.log("*****************MIDDLEWARE*****************");
 
     next();
   }
