+++ backend/src/users/decorators/update-user.decorator.ts
@@ -11,13 +11,13 @@ import {
   UserApiOkResponse,
   UserNotFoundResponse,
   UserUnprocessableEntity,
-} from "../swagger/user.swagger";
+} from "../swaggers/user.swagger";
 import {
   USER_CONFLICT_RESPONSE,
   USER_NOT_FOUND,
   USER_OK_RESPONSE,
   USER_UNPROCESSABLE_ENTITY,
-} from "src/common/constants/user.constants";
+} from "src/users/constants/user.constants";
 
 export const CustomUserUpdate = (): MethodDecorator => {
   return (
