+++ backend/src/users/entities/user.entity.ts
@@ -5,7 +5,7 @@ import { BeforeInsert, BeforeUpdate, Column, Entity } from "typeorm";
 import * as bcrypt from "bcryptjs";
 import { UnprocessableEntityException } from "@nestjs/common";
 import { SALT_ROUNDS } from "src/common/constants/core.constants";
-import { USER_UNPROCESSABLE_ENTITY } from "src/common/constants/user.constants";
+import { USER_UNPROCESSABLE_ENTITY } from "src/users/constants/user.constants";
 
 @Entity()
 export class Users extends CoreEntity {
