+++ backend/src/users/users.controller.ts
@@ -5,18 +5,21 @@ import {
   Patch,
   Param,
   ParseIntPipe,
+  UseInterceptors,
 } from "@nestjs/common";
 import { UsersService } from "./users.service";
-import { CreateUserInput, CreateUserOutput } from "./dto/create-user.dto";
-import { UpdateUserInput, UpdateUserOutput } from "./dto/update-user.dto";
+import { CreateUserInput, CreateUserOutput } from "./dtos/create-user.dto";
+import { UpdateUserInput, UpdateUserOutput } from "./dtos/update-user.dto";
 import { ApiExtraModels, ApiTags } from "@nestjs/swagger";
 import { VERSION_SWAGGER } from "src/common/constants/core.constants";
 import { CustomUserCreate } from "src/users/decorators/create-user.decorators";
 import { CustomUserUpdate } from "src/users/decorators/update-user.decorator";
 import { Users } from "./entities/user.entity";
+import { UndefinedToNullInterceptor } from "src/Interceptors/undefinedToNull.interceptor";
 
 @ApiExtraModels(Users)
 @ApiTags("Users")
+@UseInterceptors(UndefinedToNullInterceptor)
 @Controller(`api/v${VERSION_SWAGGER}/users`)
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
@@ -26,6 +29,7 @@ export class UsersController {
   createUser(
     @Body() createUserInput: CreateUserInput
   ): Promise<CreateUserOutput> {
+    console.log("*****************CONTROLLER*****************");
     return this.usersService.createUser(createUserInput);
   }
 
@@ -35,6 +39,7 @@ export class UsersController {
     @Param("id", ParseIntPipe) id: number,
     @Body() updateUserInput: UpdateUserInput
   ): Promise<UpdateUserOutput> {
+    console.log("*****************CONTROLLER*****************");
     return this.usersService.updateUser(id, updateUserInput);
   }
 }
